pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "garmin_graphics_generator"
version = "0.1.0"
description = "A CLI tool to generate hero images from watch photos."
authors = [{name = "Waclaw Kusnierczyk", email = "waclaw.kusnierczyk@gmail.com"}]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=10.0.0",
    "rembg>=2.0.50",
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0"
]

[project.scripts]
garmin-graphics-generator = "garmin_graphics_generator.cli:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]
dev = [
    "pylint",
    "tox"
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]